+++ synchroSaiso-sb/src/main/java/fr/synchoSaiso/demo/entities/Actif.java
@@ -33,7 +33,7 @@ public class Actif {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 //	private Integer idActifs;
-	private Integer actifId;
+	private Long actifId;
 	private String nameActifs;
 	private String adresseActifs;
 	private int cpActifs;
@@ -61,47 +61,6 @@ public class Actif {
 	@JsonManagedReference
 	@ManyToMany(mappedBy = "actifs")
 	private List<User> users = new ArrayList<>();
-	
-	
-//	//changement bdd
-//	@OneToMany(
-//			mappedBy = "actif",
-//			cascade = CascadeType.ALL,
-//			orphanRemoval = true
-//			)
-//	@JsonManagedReference
-//	private List<UsersActifs> users = new ArrayList<>();
-	//private Set<UsersActifs> uActifs;
-
-//	@Override
-//	public int hashCode() {
-//		return Objects.hash(actifId, adresseActifs, cpActifs, latitudeActifs, longitudeActifs, nameActifs,
-//				numeroFiscalActifs, rentcalendars, statutFiscalActifs, urlAirBnb, urlBooking, urlHomeAway,
-//				urlTripAdvisor, users, villeActifs);
-//	}
-//
-//	@Override
-//	public boolean equals(Object obj) {
-//		if (this == obj)
-//			return true;
-//		if (obj == null)
-//			return false;
-//		if (getClass() != obj.getClass())
-//			return false;
-//		Actif other = (Actif) obj;
-//		return Objects.equals(actifId, other.actifId) && Objects.equals(adresseActifs, other.adresseActifs)
-//				&& cpActifs == other.cpActifs && Objects.equals(latitudeActifs, other.latitudeActifs)
-//				&& Objects.equals(longitudeActifs, other.longitudeActifs)
-//				&& Objects.equals(nameActifs, other.nameActifs)
-//				&& Objects.equals(numeroFiscalActifs, other.numeroFiscalActifs)
-//				&& Objects.equals(rentcalendars, other.rentcalendars)
-//				&& Objects.equals(statutFiscalActifs, other.statutFiscalActifs)
-//				&& Objects.equals(urlAirBnb, other.urlAirBnb) && Objects.equals(urlBooking, other.urlBooking)
-//				&& Objects.equals(urlHomeAway, other.urlHomeAway)
-//				&& Objects.equals(urlTripAdvisor, other.urlTripAdvisor) && Objects.equals(users, other.users)
-//				&& Objects.equals(villeActifs, other.villeActifs);
-//	}
-
 	
 	
 }
